+++ packages/web3-providers/src/NextID/helpers.ts
